refactor into a few modules
